move data sources next to repo
