+++ src/packages/settings/data-types/repository/copy/data-type-copy.repository.ts
@@ -1,5 +1,5 @@
 import { UmbDataTypeRepositoryBase } from '../data-type-repository-base.js';
-import { UmbDataTypeCopyServerDataSource } from '../sources/data-type-copy.server.data.js';
+import { UmbDataTypeCopyServerDataSource } from './data-type-copy.server.data.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbCopyDataSource, UmbCopyRepository } from '@umbraco-cms/backoffice/repository';
